==> 3- CTD_PPS.ipynb <==
# Protein-protein similarity
- R package for similarity
https://cran.r-project.org/web/packages/protr/protr.pdf
- Main paper: https://bioconductor.org/packages/release/bioc/html/Biostrings.html

library(protr)
library(dplyr)
library(GOSemSim)
library(Biostrings)
library(foreach)
library(doParallel)

## Finding "uniprot" id

protlist_temp <- read.csv(file = "selected_uniport_seq.csv")
head(protlist_temp,10)

## Measuring similarity by Sequence Similarity


protlist <- c(dplyr::pull(protlist_temp,Sequence))

PPS_matrix = parSeqSim(protlist, cores=8 , batches=4, type = "local", submat = "BLOSUM62")
